+++ src/jdf/io/BASE64Encoder.cpp
@@ -1,8 +1,8 @@
 /*
- * The CIP4 Software License, Version 0.1
+ * The CIP4 Software License, Version 1.0
  *
  *
- * Copyright (c) 2001 The International Cooperation for the Integration of 
+ * Copyright (c) 2001-2012 The International Cooperation for the Integration of 
  * Processes in  Prepress, Press and Postpress (CIP4).  All rights 
  * reserved.
  *
@@ -134,17 +134,17 @@ BASE64Encoder::BASE64Encoder()
 {
 }
 
-int BASE64Encoder::bytesPerAtom() 
+size_t BASE64Encoder::bytesPerAtom() 
 {
 	return (3);
 }
 
-int BASE64Encoder::bytesPerLine() 
+size_t BASE64Encoder::bytesPerLine() 
 {
 	return (57);
 }
 
-void BASE64Encoder::encodeAtom(OutputStream& outStream, char* data, int datalen, int offset, int len) 
+void BASE64Encoder::encodeAtom(OutputStream& outStream, char* data, size_t datalen, size_t offset, size_t len) 
 {
 	char a, b, c;
 	char outBuf[4];
